Removed notes, add a stump README, moved query_api into utils.py, used HTTPAdapter for Retries
