+++ scraper.py
@@ -3,15 +3,12 @@
 from datetime import datetime
 import pandas as pd
 import requests
-import sys
-import time
 from copy import copy
 
 from config import BOARDS, DATASETS, DATA_PATH, SRC_API_URL, PT_ID
 from enrich_data import enrich_categories, enrich_levels, enrich_runs
+from utils import query_api
 
-# Length of time in seconds to sleep after each API call
-SLEEP_INTERVAL = 0.4
 
 """Loading different datasets"""
 
@@ -53,7 +50,9 @@ def get_all_runs(board_ids):
 
 
 def get_leaderboards():
-    """Get the current leaderboards for Any%, True Ending, 100%, and 101%"""
+    """Get the current leaderboards for Any%, True Ending, 100%, and 101%
+    
+    This method is currently unupdated for the new versions of the app"""
     categories = ["Any", "True_Ending", "100", "101",]
 
     current_date = datetime.utcnow().strftime("%Y-%m-%d")
@@ -110,65 +109,3 @@ def load_data(dataset, enrich_data_fun, board="PT", save_results=True):
         data_df.to_parquet(path=DATA_PATH / f"{board}_{dataset.local_path}.parquet")
 
     return data_df
-
-
-# Pagination gives you ~20 results, so you gotta call the API again using
-# the return value's pagination.links.uri for rel = 'next'
-# It ends when the pagination list doesn't have a 'next' value anymore
-def query_api(endpoint, game_id="", arg_dict=None):
-    """Query the SRC API and return all unpaginated results for endpoint with args.
-    
-    game_id is parameterized so that different games can be pulled with the same config"""
-
-    print("")
-    formatted_endpoint = endpoint.format(game_id=game_id)
-
-    # SRC results can either be paginated or unpaginated. Handle the first call, and then if
-    # it contains a `pagination` key, go into the paginator workflow
-    response = requests.get(f"{SRC_API_URL}/{formatted_endpoint}", params=arg_dict)
-    time.sleep(SLEEP_INTERVAL)
-    if response.status_code != 200:
-        print(response.text)
-        response.raise_for_status()
-
-    if "pagination" not in response.json():
-        return results['data']
-
-    # Unroll pagination to return a single list of all results
-    results_list = []
-    call_count = 1
-    while response is not None:
-        if response.status_code != 200:
-            # So far the only error I've seen for valid requests is rate limits, so
-            # sleep and retry when it happens
-            print(response.text)
-            time.sleep(25)
-            results = requests.get(next_url) if next_url else None
-            continue
-
-        results = response.json()
-        results_list += results["data"]
-
-        # Update result counts without printing a billion lines, lol
-        sys.stdout.write('\r')
-        sys.stdout.write(f"Got {len(results_list)} results")
-        sys.stdout.flush()
-
-        next_url = get_next_uri(results["pagination"])
-        response = requests.get(next_url) if next_url else None
-
-        time.sleep(SLEEP_INTERVAL)
-        call_count += 1
-
-    print("")
-
-    return results_list
-
-
-def get_next_uri(pagination_dict):
-    """Parse out the uri of the next paginated response"""
-    for link in pagination_dict.get('links', []):
-        if link['rel'] == 'next':
-            return link['uri']
-
-    return None
